[ADD] Implement ranking components: Podium, ProgressCard, RankingFilters, and RankingList; refactor Ranking component for improved structure and readability
